I2C working

+++ code/main_board/src/peripherals/ld06_board.h
@@ -18,5 +18,5 @@ esp_err_t init_ld06_board(void);
 
 esp_err_t set_cone(float center_angle, float half_cone_width);
 esp_err_t has_obstacle(bool *obstacle);
-esp_err_t closest_obstacle(float *distance);
+esp_err_t closest_obstacle(uint16_t *distance);
 //esp_err_t read_all_distances(float *distances);
